airflow/requirements.txt
requests>=2.32.0
weaviate-client>=4.6.4
psycopg2-binary>=2.9.9
neo4j>=5.22.0
boto3>=1.34.0
